requirements.txt
faiss-gpu==1.7.1
joblib==1.0.1
numpy==1.21
Pillow==9.0.0
scikit-learn==0.24.2
scipy==1.6.3
sklearn==0.0
threadpoolctl==2.1.0
torch==1.13.0+cu117
tqdm==4.61.0
typing-extensions==3.10.0.0
matplotlib==3.2.1
pandas==1.3.5
fair-esm==0.4.0
simple-parsing==0.0.19.post1
psutil==5.9.0
notifiers==1.3.0
biopython==1.79
hydra-core==1.2.0
nltk==3.7.0
